+++ packages/components/condition/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-condition@0.17.8...@synerise/ds-condition@0.17.9) (2022-06-29)
+
+**Note:** Version bump only for package @synerise/ds-condition
+
+
+
+
+
 ## [0.17.8](https://github.com/Synerise/synerise-design/compare/@synerise/ds-condition@0.17.7...@synerise/ds-condition@0.17.8) (2022-06-23)
 
 **Note:** Version bump only for package @synerise/ds-condition